ci(esp32): compile the bench and twin harnesses, and type the host-fallback macros
The two component syntax-check steps cover components/expanse/src only, so
bench_esp32_memtable.c and twin_containers.c were outside every compile lane
on every target. That is why the literal-500 divisor in the bench harness
survived: nothing built the file.

Adds a native-target lane for both TUs at -Wall -Wextra -Werror. They take
the non-ESP_PLATFORM fallback path, so the native target is the one that can
build them; test_expanse.c stays out because it needs ESP-IDF's unity.h.

-Werror required typing the two host-fallback heap macros as size_t: both
feed %zu in report_json / report_skipped, and the ESP_PLATFORM branch already
yields size_t. Those were the only two warnings in either TU.

Repoints the pending-re-measurement note at #676, the device-harvest tracker.

Refs #614, #676

+++ components/expanse/test/bench_esp32_memtable.c
@@ -34,8 +34,10 @@ static inline uint32_t get_dummy_cycles(void) {
     return (uint32_t)(ts.tv_nsec);
 }
 #define GET_CYCLES() get_dummy_cycles()
-#define GET_FREE_HEAP() (1024 * 1024)
-#define GET_LARGEST_BLOCK() (1024 * 1024)
+/* size_t, not int: both feed `%zu` in report_json / report_skipped, and the
+ * ESP_PLATFORM branch above yields size_t. */
+#define GET_FREE_HEAP() ((size_t)(1024 * 1024))
+#define GET_LARGEST_BLOCK() ((size_t)(1024 * 1024))
 #endif
 
 #define WARMUP_RUNS 3
